refactor: modifico la nueva ruta para los assets
