+++ src/components/MenuPlanets.tsx
@@ -1,8 +1,5 @@
-import { NavBarButtons } from "./NavBarButtons";
-import { NavHeader } from "./NavHeader";
-import { Line } from "./Line";
 import { useMenu } from "../hooks";
-import { ImageIcon } from "./ImageIcon";
+import { NavBarButtons, NavHeader, Line, ImageIcon } from "./";
 
 export const MenuPlanets = () => {
   const { showMenu, toggleMenu, hiddenMenu } = useMenu();
@@ -14,7 +11,7 @@ export const MenuPlanets = () => {
           the planets
         </h1>
         <ImageIcon
-          src={`/icon-hamburger-${showMenu ? "open" : "close"}.svg`}
+          src={`/assets/icon-hamburger-${showMenu ? "open" : "close"}.svg`}
           className="cursor-pointer w-6 h-4 md:hidden"
           altImage="icono de menú para celulares"
           eventClick={toggleMenu}
